Refactor import statements to use new task and workflow modules
