+++ examples/github_stars.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 import httpx
 
 from flux import ExecutionContext
-from flux import task
-from flux import workflow
+from flux.task import task
+from flux.workflow import workflow
 
 
 @task
