+++ examples/pause.py
@@ -2,10 +2,11 @@ from __future__ import annotations
 
 from datetime import timedelta
 
-from flux import ExecutionContext, workflow
-from flux.decorators import task
+from flux import ExecutionContext
+from flux.task import task
 from flux.tasks import pause
 from flux.tasks import sleep
+from flux.workflow import workflow
 
 
 @task
