+++ examples/subflows.py
@@ -4,8 +4,8 @@ import httpx
 
 from flux import call
 from flux import ExecutionContext
-from flux import task
-from flux import workflow
+from flux.task import task
+from flux.workflow import workflow
 
 
 @task
