+++ flux/__init__.py
@@ -7,12 +7,12 @@ from flux.logging import configure_logging, get_logger
 configure_logging()
 
 # First import the core domain classes to avoid circular imports
+from flux.domain.events import ExecutionEvent, ExecutionEventType, ExecutionState
 from flux.domain.execution_context import ExecutionContext
-from flux.domain.events import *
 
 # Then import the rest of the modules
-from flux.decorators import task
-from flux.decorators import workflow
+from flux.task import task, TaskMetadata
+from flux.workflow import workflow
 from flux.encoders import *
 from flux.output_storage import *
 from flux.secret_managers import *
@@ -25,5 +25,9 @@ logger = get_logger("flux")
 __all__ = [
     "task",
     "workflow",
+    "TaskMetadata",
+    "ExecutionEvent",
+    "ExecutionState",
+    "ExecutionEventType",
     "ExecutionContext",
 ]
