+++ tests/flux/test_output_storage_integration.py
@@ -5,8 +5,8 @@ from unittest.mock import patch
 import pytest
 
 from flux import ExecutionContext
-from flux.decorators import task
-from flux.decorators import workflow
+from flux.task import task
+from flux.workflow import workflow
 from flux.output_storage import InlineOutputStorage
 from flux.output_storage import OutputStorageReference
